+++ src/artists/artists.controller.ts
@@ -6,7 +6,9 @@ import {
   NotFoundException,
   Param,
   Post,
+  Req,
   UploadedFile,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
@@ -14,6 +16,9 @@ import { Artist, ArtistDocument } from '../schemas/artist.schema';
 import { Model } from 'mongoose';
 import { CreateArtistDto } from './create-artist.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
+import { TokenAuthGuard } from '../token-auth/token-auth.guard';
+import { Request } from 'express';
+import { IUserDocument, User } from '../schemas/user.schema';
 
 @Controller('artists')
 export class ArtistsController {
@@ -22,6 +27,7 @@ export class ArtistsController {
     private artistModel: Model<ArtistDocument>,
   ) {}
 
+  @UseGuards(TokenAuthGuard)
   @Get()
   getAll() {
     return this.artistModel.find();
@@ -32,6 +38,7 @@ export class ArtistsController {
     return this.artistModel.findById(id);
   }
 
+  @UseGuards(TokenAuthGuard)
   @Post()
   @UseInterceptors(FileInterceptor('image', { dest: './public/uploads/artist' }))
   async create(@UploadedFile() file: Express.Multer.File, @Body() artistDto: CreateArtistDto) {
@@ -44,10 +51,14 @@ export class ArtistsController {
     return artist;
   }
 
+  @UseGuards(TokenAuthGuard)
   @Delete(':id')
-  async deleteOne(@Param('id') id: string) {
-    const artist = await this.artistModel.findOneAndDelete({ _id: id });
-    if (!artist) throw new NotFoundException(`Artist with id ${id} not found`);
-    return { message: 'Artists deleted successfully', artist };
+  async deleteOne(@Param('id') id: string, @Req() req: Request<{ user: User }>) {
+    const user = req.user as IUserDocument;
+    if (user.role === 'admin') {
+      const artist = await this.artistModel.findOneAndDelete({ _id: id });
+      if (!artist) throw new NotFoundException(`Artist with id ${id} not found`);
+      return { message: 'Artists deleted successfully', artist };
+    }
   }
 }
